Rewritten Anime, Fun, Help cogs and their respective utils
