+++ cogs/anime.py
@@ -11,15 +11,24 @@ from bot import Bot
 
 
 class Anime(commands.Cog):
-    """
-    For Anime fun
-    """
+    """For Anime fun"""
 
     def __init__(self, bot: Bot):
         self.bot = bot
 
     @staticmethod
     def neko_api(ctx, x, msg=""):
+        """
+        Fetches data from the nekos.fun API and creates an embed with the image.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - x (str): The type of neko image to fetch.
+        - msg (str): Additional message to include in the embed.
+
+        Returns:
+        - embed (disnake.Embed): The embed with the neko image.
+        """
         req = requests.get(f"http://api.nekos.fun:8080/api/{x}")
         if req.status_code != 200:
             raise NoNeko("Couldn't load image")
@@ -38,13 +47,23 @@ class Anime(commands.Cog):
     async def info(
         self, ctx: disnake.AppCmdInter, name=Param(description="Title of the series")
     ):
+        """
+        Retrieves information about an anime or manga series and sends an embed with the details.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - name (str): The title of the anime or manga series.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         anime = animec.Anime(name)
         try:
             embed = disnake.Embed(
                 title=anime.title_english,
                 url=anime.url,
-                description=f"{anime.description[:300]}.....",
+                description=f"{anime.description[:300]}....",
                 color=disnake.Colour.random(),
             )
             embed.add_field(name="Episodes", value=str(anime.episodes))
@@ -66,7 +85,7 @@ class Anime(commands.Cog):
             embed.add_field(name="Genres", value=", ".join(anime.genres))
             embed.add_field(name="Teaser", value=str(anime.teaser))
             embed.add_field(name="Recommend", value=", ".join(anime.recommend()))
-            embed.add_field(name="NSFW status", value=str(anime.is_nsfw()))
+            embed.add_field(name="NSFW status", value="🔞" if anime.is_nsfw() else "🆗")
             embed.set_thumbnail(url=anime.poster)
             embed.set_author(
                 name=ctx.bot.user.display_name, icon_url=f"{ctx.bot.user.avatar.url}"
@@ -92,8 +111,20 @@ class Anime(commands.Cog):
         name="anime-character", description="Get image of an Anime or Manga character"
     )
     async def character(
-        self, ctx: disnake.AppCmdInter, name=Param(description="Name of the character")
+        self,
+        ctx: disnake.AppCmdInter,
+        name: str = Param(description="Name of the character"),
     ):
+        """
+        Retrieves the image of an anime or manga character and sends an embed with the image.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - name (str): The name of the character.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         try:
             char = animec.Charsearch(name)
@@ -124,6 +155,17 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         amount: int = Param(3, description="No of news per request"),
     ):
+        """
+        Retrieves the latest anime news and sends an embed with the news articles.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - amount (int): The number of news articles to retrieve.
+
+        Returns:
+        - None
+        """
+        num_emojis = ["1️⃣", "2️⃣", "3️⃣", "4️⃣", "5️⃣", "6️⃣", "7️⃣", "8️⃣"]
         await ctx.response.defer()
         try:
             news = animec.Aninews(amount)
@@ -141,7 +183,7 @@ class Anime(commands.Cog):
         descriptions = news.description
 
         embed = disnake.Embed(
-            title="Latest Anime News",
+            title="Latest Anime News 🗞️",
             color=disnake.Colour.random(),
             timestamp=datetime.datetime.utcnow(),
         )
@@ -156,8 +198,8 @@ class Anime(commands.Cog):
 
         for i in range(amount):
             embed.add_field(
-                name=f"{i+1}) {titles[i]}",
-                value=f"{descriptions[i][:200]}...\n[Read more]({links[i]})",
+                name=f"{num_emojis[i]} {titles[i]}",
+                value=f"> {descriptions[i][:200]}...\n[Read more]({links[i]})",
                 inline=False,
             )
         await ctx.send(embed=embed)
@@ -166,6 +208,15 @@ class Anime(commands.Cog):
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.slash_command(name="emote-neko", description="Post a neko")
     async def neko(self, ctx: disnake.AppCmdInter):
+        """
+        Sends an embed with a random neko image.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         await ctx.send(embed=self.neko_api(ctx, "neko"))
 
@@ -173,6 +224,15 @@ class Anime(commands.Cog):
     @commands.cooldown(1, 5)
     @commands.slash_command(name="emote-waifu", description="Post a waifu")
     async def waifu(self, ctx: disnake.AppCmdInter):
+        """
+        Sends an embed with a random waifu image.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         await ctx.send(embed=self.neko_api(ctx, "waifu"))
 
@@ -184,6 +244,16 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         member: disnake.Member = Param(description="mention a member"),
     ):
+        """
+        Sends an embed with a neko image and mentions the member being tickled.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - member (disnake.Member): The member to be tickled.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         try:
             await ctx.send(
@@ -202,6 +272,16 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         member: disnake.Member = Param(description="mention a member"),
     ):
+        """
+        Sends an embed with a neko image and mentions the member being poked.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - member (disnake.Member): The member to be poked.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         try:
             await ctx.send(
@@ -218,6 +298,16 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         member: disnake.Member = Param(description="mention a member"),
     ):
+        """
+        Sends an embed with a neko image and mentions the members involved in a kiss.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - member (disnake.Member): The member to be kissed.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         try:
             await ctx.send(
@@ -236,6 +326,16 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         member: disnake.Member = Param(description="mention a member"),
     ):
+        """
+        Sends an embed with a neko image and mentions the member being slapped.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - member (disnake.Member): The member to be slapped.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         try:
             await ctx.send(
@@ -254,6 +354,16 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         member: disnake.Member = Param(description="mention a member"),
     ):
+        """
+        Licks a specified member.
+
+        Parameters:
+        - ctx: The context of the command.
+        - member: The member to be licked.
+
+        Returns:
+        None
+        """
         await ctx.response.defer()
         try:
             await ctx.send(embed=self.neko_api(ctx, "lick", f"Licked {member.mention}"))
@@ -268,6 +378,16 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         member: disnake.Member = Param(description="mention a member"),
     ):
+        """
+        Sends a hug message to the specified member.
+
+        Parameters:
+        - ctx: The context of the command.
+        - member: The member to hug.
+
+        Returns:
+        None
+        """
         await ctx.response.defer()
         try:
             await ctx.send(embed=self.neko_api(ctx, "hug", f"Hugged {member.mention}"))
@@ -282,6 +402,16 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         member: disnake.Member = Param(description="mention a member"),
     ):
+        """
+        Pat a member.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context of the command.
+        - member (disnake.Member): The member to pat.
+
+        Returns:
+        None
+        """
         await ctx.response.defer()
         try:
             await ctx.send(embed=self.neko_api(ctx, "pat", f"patted {member.mention}"))
@@ -292,6 +422,15 @@ class Anime(commands.Cog):
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.slash_command(name="emote-cry", description="Show you're broken")
     async def cry(self, ctx: disnake.AppCmdInter):
+        """
+        Sends a cry animation.
+
+        Parameters:
+        - ctx: The context object representing the command invocation.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         await ctx.send(embed=self.neko_api(ctx, "cry"))
 
@@ -299,6 +438,15 @@ class Anime(commands.Cog):
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.slash_command(name="emote-smug", description="Smug")
     async def smug(self, ctx: disnake.AppCmdInter):
+        """
+        Sends a smug anime image.
+
+        Parameters:
+        - ctx: The Discord context object.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         await ctx.send(embed=self.neko_api(ctx, "smug"))
 
@@ -306,6 +454,15 @@ class Anime(commands.Cog):
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.slash_command(name="emote-lol", description="Laughing out loud")
     async def lol(self, ctx: disnake.AppCmdInter):
+        """
+        Sends a laughing image.
+
+        Parameters:
+        - ctx (disnake.AppCmdInter): The context object representing the command invocation.
+
+        Returns:
+        - None
+        """
         await ctx.response.defer()
         await ctx.send(embed=self.neko_api(ctx, "laugh"))
 
@@ -317,6 +474,16 @@ class Anime(commands.Cog):
         ctx: disnake.AppCmdInter,
         member: disnake.Member = Param(description="mention a member"),
     ):
+        """
+        Feed a member.
+
+        Parameters:
+        - ctx: The context of the command.
+        - member: The member to feed.
+
+        Returns:
+        None
+        """
         await ctx.response.defer()
         try:
             await ctx.send(embed=self.neko_api(ctx, "feed", f"fed {member.mention}"))
