+++ cogs/nsfw.py
@@ -34,67 +34,67 @@ class Nsfw(commands.Cog):
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="fourk", description="4K", nsfw=True)
+    @commands.slash_command(name="nsfw-fourk", description="4K", nsfw=True)
     async def fourk(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "4k"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="ass", description="A sexy ass", nsfw=True)
+    @commands.slash_command(name="nsfw-ass", description="A sexy ass", nsfw=True)
     async def ass(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "ass"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="blowjob", description="For a blowjob", nsfw=True)
+    @commands.slash_command(name="nsfw-blowjob", description="For a blowjob", nsfw=True)
     async def blowjob(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "blowjob"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="boobs", description="Boobies", nsfw=True)
+    @commands.slash_command(name="nsfw-boobs", description="Boobies", nsfw=True)
     async def boobs(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "boobs"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="cum", description="The baby-gravy", nsfw=True)
+    @commands.slash_command(name="nsfw-cum", description="The baby-gravy", nsfw=True)
     async def cum(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "cum"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="feet", description="bottom view...", nsfw=True)
+    @commands.slash_command(name="nsfw-feet", description="bottom view...", nsfw=True)
     async def feet(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "feet"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="hentai", description="Random hentai", nsfw=True)
+    @commands.slash_command(name="nsfw-hentai", description="Random hentai", nsfw=True)
     async def hentai(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "hentai"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="spank", description="booty spank", nsfw=True)
+    @commands.slash_command(name="nsfw-spank", description="booty spank", nsfw=True)
     async def spank(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "spank"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="gasm", description="Or-Gasm", nsfw=True)
+    @commands.slash_command(name="nsfw-gasm", description="Or-Gasm", nsfw=True)
     async def gasm(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "gasm"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="lesbian", description="lesbian fuck", nsfw=True)
+    @commands.slash_command(name="nsfw-lesbian", description="lesbian fuck", nsfw=True)
     async def lesbian(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "lesbian"), ephemeral=True)
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    @commands.slash_command(name="pussy", description="Pussy, not a cat", nsfw=True)
+    @commands.slash_command(name="nsfw-pussy", description="Pussy, not a cat", nsfw=True)
     async def pussy(self, ctx: disnake.AppCmdInter):
         await ctx.send(embed=self.neko_api(ctx, "pussy"), ephemeral=True)
