+++ utils/funHelpers/__init__.py
@@ -1,0 +1,5 @@
+from utils.funHelpers.mineswiper import MineswiperView
+from utils.funHelpers.tictactoe import TicTacToe
+from utils.funHelpers.rps import RPS
+from utils.funHelpers.misc import *
+from utils.funHelpers.akinator import AkinatorGame
